# swagger/swagger.json
{"swagger":"2.0","info":{"title":"goa-sandbox","description":"sample","contact":{"name":"tkc","email":"iijima@onga.jp","url":"https://github.com/tkc/goa-sandbox/issues"},"license":{"name":"MIT","url":"https://github.com/tkc/goa-sandbox/blob/master/LICENSE"},"version":""},"host":"localhost:8080","basePath":"/api/v1","schemes":["http"],"consumes":["application/json","application/xml","application/gob","application/x-gob"],"produces":["application/json","application/xml","application/gob","application/x-gob"],"paths":{"/currentuser":{"get":{"tags":["account"],"summary":"currentUser account","operationId":"account#currentUser","produces":["application/vnd.goa.error","application/vnd.goa.example.account.currentuser+json"],"parameters":[{"name":"token","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/GoaExampleAccountCurrentuser"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}},"/jwt":{"get":{"tags":["jwt"],"summary":"secure jwt","description":"This action is secured with the jwt scheme\n\nRequired security scopes:\n  * `api:access`","operationId":"jwt#secure","produces":["application/vnd.goa.error","application/vnd.goa.example.account.ok+json"],"parameters":[{"name":"fail","in":"query","description":"Force auth failure via JWT validation middleware","required":false,"type":"boolean"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/GoaExampleAccountOk"}},"401":{"description":"Unauthorized","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"],"security":[{"jwt":["api:access"]}]}},"/jwt/signin":{"post":{"tags":["jwt"],"summary":"signin jwt","description":"Creates a valid JWT","operationId":"jwt#signin","produces":["application/vnd.goa.error","application/vnd.goa.example.account.register+json"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/GoaExampleAccountRegister"}},"204":{"description":"No Content","headers":{"Authorization":{"description":"Generated JWT","type":"string"}}},"401":{"description":"Unauthorized","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"],"security":[{"SigninBasicAuth":[]}]}},"/jwt/unsecure":{"get":{"tags":["jwt"],"summary":"unsecure jwt","description":"This action does not require auth","operationId":"jwt#unsecure","produces":["application/vnd.goa.error","application/vnd.goa.example.account.ok+json"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/GoaExampleAccountOk"}},"401":{"description":"Unauthorized","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}},"/login":{"post":{"tags":["account"],"summary":"login account","operationId":"account#login","produces":["application/vnd.goa.error","application/vnd.goa.example.account.login+json"],"parameters":[{"name":"email","in":"query","required":false,"type":"string"},{"name":"passWord","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/GoaExampleAccountLogin"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}},"/logout":{"post":{"tags":["account"],"summary":"logout account","operationId":"account#logout","produces":["application/vnd.goa.error","application/vnd.goa.example.account.logout+json"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/GoaExampleAccountLogout"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}},"/register":{"post":{"tags":["account"],"summary":"register account","operationId":"account#register","produces":["application/vnd.goa.error","application/vnd.goa.example.account.register+json"],"parameters":[{"name":"email","in":"query","required":false,"type":"string"},{"name":"name","in":"query","required":false,"type":"string"},{"name":"passWord","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/GoaExampleAccountRegister"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}}},"definitions":{"GoaExampleAccountCurrentuser":{"title":"Mediatype identifier: application/vnd.goa.example.account.currentuser+json; view=default","type":"object","properties":{"name":{"type":"string","description":"Name of user","example":"Laborum voluptas a officiis."},"userId":{"type":"integer","description":"Unique User ID","example":1970807810636653972,"format":"int64"}},"description":"GoaExampleAccountCurrentuser media type (default view)","example":{"name":"Laborum voluptas a officiis.","userId":1970807810636653972},"required":["userId","name"]},"GoaExampleAccountLogin":{"title":"Mediatype identifier: application/vnd.goa.example.account.login+json; view=default","type":"object","properties":{"name":{"type":"string","description":"Name of user","example":"Eum molestiae temporibus et."},"token":{"type":"string","description":"Unique User Token","example":"Voluptatem repellendus."},"userId":{"type":"integer","description":"Unique User ID","example":4887016811865582828,"format":"int64"}},"description":"GoaExampleAccountLogin media type (default view)","example":{"name":"Eum molestiae temporibus et.","token":"Voluptatem repellendus.","userId":4887016811865582828},"required":["userId","token","name"]},"GoaExampleAccountLogout":{"title":"Mediatype identifier: application/vnd.goa.example.account.logout+json; view=default","type":"object","properties":{"status":{"type":"boolean","description":"action status","example":false}},"description":"GoaExampleAccountLogout media type (default view)","example":{"status":false}},"GoaExampleAccountOk":{"title":"Mediatype identifier: application/vnd.goa.example.account.ok+json; view=default","type":"object","properties":{"status":{"type":"boolean","description":"action status","example":false}},"description":"GoaExampleAccountOk media type (default view)","example":{"status":false}},"GoaExampleAccountRegister":{"title":"Mediatype identifier: application/vnd.goa.example.account.register+json; view=default","type":"object","properties":{"name":{"type":"string","description":"Name of user","example":"Velit debitis distinctio velit."},"token":{"type":"string","description":"Unique User Token","example":"Iste odio explicabo quae accusantium id et."},"userId":{"type":"integer","description":"Unique User ID","example":8093784523514649816,"format":"int64"}},"description":"GoaExampleAccountRegister media type (default view)","example":{"name":"Velit debitis distinctio velit.","token":"Iste odio explicabo quae accusantium id et.","userId":8093784523514649816},"required":["userId","token","name"]},"error":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"code":{"type":"string","description":"an application-specific error code, expressed as a string value.","example":"invalid_value"},"detail":{"type":"string","description":"a human-readable explanation specific to this occurrence of the problem.","example":"Value of ID must be an integer"},"id":{"type":"string","description":"a unique identifier for this particular occurrence of the problem.","example":"3F1FKVRR"},"meta":{"type":"object","description":"a meta object containing non-standard meta-information about the error.","example":{"timestamp":1458609066},"additionalProperties":true},"status":{"type":"string","description":"the HTTP status code applicable to this problem, expressed as a string value.","example":"400"}},"description":"Error response media type (default view)","example":{"code":"invalid_value","detail":"Value of ID must be an integer","id":"3F1FKVRR","meta":{"timestamp":1458609066},"status":"400"}}},"securityDefinitions":{"SigninBasicAuth":{"type":"basic"},"jwt":{"type":"apiKey","description":"\n\n**Security Scopes**:\n  * `api:access`: API access","name":"Authorization","in":"header"}},"externalDocs":{"description":"wiki","url":"https://github.com/tkc/goa-sandbox/wiki"}}